+++ tests/test.h
@@ -3,6 +3,7 @@
 
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdlib.h>
 
 void test_run(void (*fn)());
 void test_fail();