+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unpublished
 
+* Free memory in tests and example.
 * Improved build setup (no need to list test names).
 
 ### v0.1.4 (2020-09-25)